Fix drag preview shape
Previously all blocks had a statement shape as drag preview. Now that we
have a good mapping of blocks / shapes, use the corresponding shape.

A especial case is the control block which uses the background twice,
one for the top part and one for the bottom part. For simplicity, use
the statement shape in this case.

Also, move drag colors to constants.