Set Godot 4.3 as minimum supported version
Godot 4.3 has been available for 4 months and it seems unlikely that anyone
will have a pressing need to add blocks to a Godot 4.2 project.

Although we have been running unit tests against 4.2 in CI, those do not
cover the full functionality of the plugin. To the best of my knowledge,
no-one is regularly testing the plugin interactively against Godot 4.2.

Set the minimum version in the asset library template. Drop 4.2 from the
CI test matrix.