Fix types relationships
The types relationship (previously named cast_relationships) tell which
type can be used in the context of another type. For example, integers
and floats are exchangeable. And in particular, all other types can be
used in boolean context: `if not Color.BLACK:` is allowed in GDScript as
Color.BLACK is falsey. Same for the empty string, zero number, etc.

This fixes a bug where an integer or float value block was allowed to be
snapped into a string slot.

This removes outdated custom casting code that has been abandoned.

https://github.com/endlessm/godot-block-coding/issues/351