Make blocks in Picker not editable
Add a new "editable" boolean property to Block, TemplateEditor and
ParameterInput. When the block is in the picker, it will not be
editable. Only when they are dragged to the canvas they become editable.

The property has to be in each of these components because Block can't
access its ParameterInputs directly, only through its TemplateEditor.

The dropdown options in a not-editable block can still be changed to
avoid disabling the button. But the change doesn't trigger any effect.

The default is editable = true, so blocks in the canvas are editable.

Fix https://github.com/endlessm/godot-block-coding/issues/357