Don't warn when BlockCode script extends parent script
Commit 87c31b05c9f720cc2512285340117b86afc94e8e ("Warn when Block Code
overrides parent script") introduced a warning when a BlockCode node's
parent has a script attached. But if the parent is (for example) a
Character2D with our SimpleCharacter script attached, this warning is
incorrect: the block code script will correctly extend, not override,
the parent.

Before issuing this warning, check if the parent's script has a
class_name that matches what the generated BlockCode script will extend,
and suppress the warning if so.

Fixes https://github.com/endlessm/godot-block-coding/issues/356