Update technical documentation
Document BlockDefinition resource and block types using GDScript
documentation comments.

Outside code, rename the current technical document to
`docs/TECH_OVERVIEW.md` to indicate that it's technical. And update it:
- Move "How to create a new block" to the top
- Leave UI technical docs in its own section
- Remove other outdated explanation

It would be nice to also document the AST decouple from the UI, but I'm
no expert to do it.